Safari-611.1.10/LayoutTests/http/tests/contentdispositionattachmentsandbox/scripts-disabled.html

<!DOCTYPE html>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
if (window.internals)
    internals.settings.setContentDispositionAttachmentSandboxEnabled(true);
</script>
<p>This test verifies that scripts are disabled when 'Content-Disposition: attachment' sandboxing is enabled. A security error will be logged to the console if the test passes.</p>
<iframe src="resources/scripts-frame.php"></iframe>
